refactor(route): create protected route
